# apps/management-api/openapi.yml
openapi: 3.0.3
info:
  title: Podverse Management API
  description: Comprehensive OpenAPI specification generated from audited route matrix coverage for Podverse management API.
  version: 2.0.0
servers:
  - url: https://management-api.podverse.fm/api/v2
    description: Production Server
  - url: http://localhost:3100/api/v2
    description: Local Development Server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: pv_mgmt_auth
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        code:
          type: string
          nullable: true
          description: Machine-readable error code when available.
        details:
          type: array
          nullable: true
          description: Optional field-level or contextual error details.
          items:
            $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required:
        - message
      properties:
        field:
          type: string
          nullable: true
          description: Request field associated with this detail entry when applicable.
        message:
          type: string
          description: Human-readable detail message.
        code:
          type: string
          nullable: true
          description: Optional machine-readable detail code.
    ValidationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            message:
              type: string
              description: Error description
              example: Validation failed
            code:
              type: string
              example: validation_error
    AuthenticationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            message:
              type: string
              description: Authentication failure reason
              example: Unauthorized
            code:
              type: string
              example: unauthorized
    AuthorizationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            message:
              type: string
              description: Authorization failure reason
              example: Forbidden
            code:
              type: string
              example: forbidden
    NotFoundErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            message:
              type: string
              description: Resource not found
              example: Not found
            code:
              type: string
              example: not_found
    InternalServerErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            message:
              type: string
              example: Internal server error
            code:
              type: string
              example: internal_server_error
...
    DatabaseMutationData:
      type: object
      minProperties: 1
      maxProperties: 20
      additionalProperties:
        anyOf:
          - type: string
          - type: number
          - type: boolean
        nullable: true
    DatabaseFieldMeta:
      type: object
      required:
        - name
        - type
        - nullable
        - updatable
      properties:
        name:
          type: string
        type:
          type: string
          enum:
            - integer
            - text
            - boolean
            - timestamp
        nullable:
          type: boolean
        updatable:
          type: boolean
    DatabaseTableMeta:
      type: object
      required:
        - tableName
        - primaryKeyField
        - fields
        - defaultSortField
        - defaultSortDirection
        - maxPageSize
        - readOnly
      properties:
        tableName:
          type: string
        primaryKeyField:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/DatabaseFieldMeta"
        defaultSortField:
          type: string
        defaultSortDirection:
          type: string
          enum:
            - ASC
            - DESC
        maxPageSize:
          type: integer
          minimum: 1
        readOnly:
          type: boolean
    DatabaseFilterCondition:
      type: object
      required:
        - field
        - operator
      properties:
        field:
          type: string
        operator:
          type: string
          enum:
            - eq
            - neq
            - gt
            - gte
            - lt
            - lte
            - like
            - in
            - is_null
            - is_not_null
        value:
          oneOf:
            - type: string
            - type: number
            - type: array
              maxItems: 50
              items:
                oneOf:
                  - type: string
                  - type: number
    DatabaseSortCondition:
      type: object
      required:
        - field
        - direction
      properties:
        field:
          type: string
        direction:
          type: string
          enum:
            - ASC
            - DESC
    ManagementPostDatabaseTableRequest:
      $ref: "#/components/schemas/DatabaseMutationData"
    ManagementPostDatabaseTableResponse201:
      type: object
      description: Operation-specific response payload for POST /database/{table} (201).
      additionalProperties: true
    ManagementDeleteDatabaseTableIdResponse200:
      type: object
      description: Operation-specific response payload for DELETE /database/{table}/{id} (200).
      additionalProperties: true
    ManagementGetDatabaseTableIdResponse200:
      type: object
      description: Operation-specific response payload for GET /database/{table}/{id} (200).
      additionalProperties: true
    ManagementPatchDatabaseTableIdRequest:
      $ref: "#/components/schemas/DatabaseMutationData"
    ManagementPatchDatabaseTableIdResponse200:
      type: object
      description: Operation-specific response payload for PATCH /database/{table}/{id} (200).
      additionalProperties: true
    ManagementGetDatabaseTableMetaResponse200:
      $ref: "#/components/schemas/DatabaseTableMeta"
    ManagementPostDatabaseTableQueryRequest:
      type: object
      description: Operation-specific request payload for POST /database/{table}/query.
      properties:
        filters:
          type: array
          maxItems: 10
          items:
            $ref: "#/components/schemas/DatabaseFilterCondition"
        sorts:
          type: array
          maxItems: 3
          items:
            $ref: "#/components/schemas/DatabaseSortCondition"
        page:
          type: integer
          minimum: 1
          default: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
          default: 25
      additionalProperties: false
    ManagementPostDatabaseTableQueryResponse200:
      type: object
      description: Operation-specific response payload for POST /database/{table}/query (200).
      required:
        - rows
        - total
      properties:
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
        total:
          type: integer
          minimum: 0
    ManagementGetDatabaseTablesResponse200:
      type: object
      description: Operation-specific response payload for GET /database/tables (200).
      required:
        - tables
      properties:
        tables:
          type: array
          items:
            $ref: "#/components/schemas/DatabaseTableMeta"
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /database/{table}:
    post:
      operationId: managementPostDatabaseTable
      tags:
        - database
      summary: POST /database/{table}
      description: "Auto-generated from route matrix. Authenticated endpoint (cookieAuth or bearerAuth). Permission/auth note: authenticated."
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagementPostDatabaseTableRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementPostDatabaseTableResponse201"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /database/{table}/{id}:
    delete:
      operationId: managementDeleteDatabaseTableId
      tags:
        - database
      summary: DELETE /database/{table}/{id}
      description: "Auto-generated from route matrix. Authenticated endpoint (cookieAuth or bearerAuth). Permission/auth note: authenticated."
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementDeleteDatabaseTableIdResponse200"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
    get:
      operationId: managementGetDatabaseTableId
      tags:
        - database
      summary: GET /database/{table}/{id}
      description: "Auto-generated from route matrix. Authenticated endpoint (cookieAuth or bearerAuth). Permission/auth note: authenticated."
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementGetDatabaseTableIdResponse200"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
    patch:
      operationId: managementPatchDatabaseTableId
      tags:
        - database
      summary: PATCH /database/{table}/{id}
      description: "Auto-generated from route matrix. Authenticated endpoint (cookieAuth or bearerAuth). Permission/auth note: authenticated."
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagementPatchDatabaseTableIdRequest"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementPatchDatabaseTableIdResponse200"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /database/{table}/meta:
    get:
      operationId: managementGetDatabaseTableMeta
      tags:
        - database
      summary: GET /database/{table}/meta
      description: "Auto-generated from route matrix. Authenticated endpoint (cookieAuth or bearerAuth). Permission/auth note: authenticated."
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementGetDatabaseTableMetaResponse200"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /database/{table}/query:
    post:
      operationId: managementPostDatabaseTableQuery
      tags:
        - database
      summary: POST /database/{table}/query
      description: "Auto-generated from route matrix. Authenticated endpoint (cookieAuth or bearerAuth). Permission/auth note: Requires CRUD permissions specific to the requested table."
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagementPostDatabaseTableQueryRequest"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementPostDatabaseTableQueryResponse200"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /database/tables:
    get:
      operationId: managementGetDatabaseTables
      tags:
        - database
      summary: GET /database/tables
      description: "Auto-generated from route matrix. Authenticated endpoint (cookieAuth or bearerAuth). Permission/auth note: authenticated."
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementGetDatabaseTablesResponse200"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
...